edited comments

+++ src/JEncoder/JEncoderAS5048bI2C.h
@@ -1,10 +1,9 @@
 /**
- * @brief reads a type of absolute encoder https://ams.com/as0548b (uses I2C)
+ * @brief reads a type of absolute encoder https://ams.com/as0548b (uses I2C) \n
  * based on work by sosandroid https://github.com/sosandroid/AMS_AS5048B 
- * @note remember to add pull up resistors on I2C lines (~4.7k for 5v, ~2.4k for 3.3v microcontrollers)
- * #define J_ENCODER_AS5048_I2C_USE_SOFTWIRE to use SoftWire
+ * @note remember to add pull up resistors on I2C lines (~4.7k for 5v, ~2.4k for 3.3v microcontrollers). \n
+ * Use #define J_ENCODER_AS5048_I2C_USE_SOFTWIRE to use SoftWire
  */
-
 #ifndef J_ENCODER_AS5048B_I2C_H
 #define J_ENCODER_AS5048B_I2C_H
 #include "JEncoder.h"
@@ -223,7 +222,7 @@ public:
      */
     void setEncoderZero(int zeroAngle)
     {
-        zeroAngle = constrain(zeroAngle, 0, STEPS_PER_TURN - 1);
+        zeroAngle = constrain(zeroAngle, 0, (long)STEPS_PER_TURN - 1);
         writeToZeroRegister(zeroAngle);
     }
